requirements.txt
APScheduler==3.10.1
cachetools==5.3.1
certifi==2023.5.7
python-telegram-bot==13.15
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
requests==2.31.0
openai~=0.27.7
psycopg2~=2.9.6
beautifulsoup4~=4.12.2
